% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{bestnet}
\alias{bestnet}
\title{Given a list of best sinks, find a best network.}
\usage{
bestnet(bsinks, m)
}
\arguments{
\item{bsinks}{A list of best sinks.}

\item{m}{Number of variables in the whole data.}
}
\value{
A data frame of a best network.
}
\description{
Given a list of best sinks, find a best network.
}
