% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{bestps}
\alias{bestps}
\title{Find the best parent set of a node, for parent set pset, given the
list of Possible parents sets' subsets, list of Possible parents sets' subsets
scores, and list of scores of all nodes with no parents.}
\usage{
bestps(v, pset, pps, ppss, ms, max_parents, surv)
}
\arguments{
\item{v}{node variable.}

\item{pset}{Parents of v.}

\item{pps}{List of Possible parents sets' subsets.}

\item{ppss}{List of Possible parents sets' subsets scores.}

\item{ms}{List of scores of all nodes with no parents.}

\item{max_parents}{Maximum in-degree.}

\item{surv}{Whether Survival outcome data, 0 or 1.}
}
\value{
The score of a node, given its parent set, and the list of parent set scores.
}
\description{
Find the best parent set of a node, for parent set pset, given the
list of Possible parents sets' subsets, list of Possible parents sets' subsets
scores, and list of scores of all nodes with no parents.
}
