% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestSink.R
\name{bestSinks}
\alias{bestSinks}
\title{Find best sinks for all subsets of variables in data.}
\usage{
bestSinks(pp, ms, po, pps, ppss, bps, mydata, surv)
}
\arguments{
\item{pp}{List of possible parent sets for all variables.}

\item{ms}{List of scores of nodes with no parents.}

\item{po}{List of possible offspring sets for all variables.}

\item{pps}{List of Possible parents sets' subsets.}

\item{ppss}{List of Possible parents sets' subsets' scores.}

\item{bps}{List of best parents sets' for each parents subsets'.}

\item{mydata}{Data frame of data.}

\item{surv}{Whether Survival outcome data, 0 or 1.}
}
\value{
A data frame of best sinks for each subset with score.
}
\description{
Find best sinks for all subsets of variables in data.
}
