% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bestSinksCnew}
\alias{bestSinksCnew}
\title{Internal function to find the Best Sink}
\usage{
bestSinksCnew(pp, ms, po, pps, ppss, bps, mydata)
}
\arguments{
\item{pp}{list}

\item{ms}{list}

\item{po}{list}

\item{pps}{list}

\item{ppss}{list}

\item{bps}{list}

\item{mydata}{list}
}
\description{
Internal function to find the Best Sink
}
