% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causnet-package.R
\docType{package}
\name{causnet-package}
\alias{causnet}
\alias{causnet-package}
\title{causnet: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/USCbiostats/causnet}
  \item Report bugs at \url{https://github.com/USCbiostats/causnet/issues}
}

}
\author{
\strong{Maintainer}: Emil Hvitfeldt \email{emilhhvitfeldt@gmail.com} (\href{https://orcid.org/0000-0002-0679-1945}{ORCID})

Authors:
\itemize{
  \item Joshua Millstein \email{joshua.millstein@usc.edu}
}

}
\keyword{internal}
