% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causnet.R
\name{causnet}
\alias{causnet}
\title{Find Globally Optimal Causal Network}
\usage{
causnet(data, alpha = 0.05)
}
\arguments{
\item{data}{A data.frame of numeric variables.}

\item{alpha}{Numerical, p-value threshold to determine cutoff for possible
parents.}
}
\value{
A list with 2 elements: a data data.frame with 3 columns; from, to
and component, and a vector of number of best parents.
}
\description{
Find Globally Optimal Causal Network
}
\examples{
new_data <- simdat(n_var = 5)
out <- causnet(data = new_data, 0.5)
}
