% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{comb1}
\alias{comb1}
\title{Create a list of all possible subsets of a given Set upto specified cardinality.}
\usage{
comb1(vec, maxSubsetSize)
}
\arguments{
\item{vec}{Vector of the Superset.}

\item{maxSubsetSize}{The maximum cardinality of the subsets of a given Superset to be found.}
}
\value{
A list of all possible subsets of a given Set upto specified cardinality.
}
\description{
Create a list of all possible subsets of a given Set upto specified cardinality.
}
