% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{get.score}
\alias{get.score}
\title{Get the score of a node, given its parent set, and the list of parent set scores.}
\usage{
get.score(v, pset, pps, ppss, ms)
}
\arguments{
\item{v}{node variable.}

\item{pset}{Parents of v.}

\item{pps}{Possible parents sets' subsets.}

\item{ms}{List of scores of nodes with no parents.}

\item{mydata}{Data frame of data with last column as output variable in case of binary/continuous outcome;
in case of survival data, data with last two columns the survival outcome and time.}
}
\value{
The score of a node, given its parent set, and the list of parent set scores.
}
\description{
Get the score of a node, given its parent set, and the list of parent set scores.
}
