% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{list_setequal}
\alias{list_setequal}
\title{Set equal for lists}
\usage{
list_setequal(x, y)
}
\arguments{
\item{x}{list of integer vectors}

\item{y}{integer vectors}
}
\description{
Returns logical vector with index of where y is equal to x.
}
\details{
It is assumed that y only appears in x once and that all vectors in x are
sorted.
}
