% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{mypp}
\alias{mypp}
\title{Find possible parents for each variable based on FDR or correlation curoff.}
\usage{
mypp(mydata, alpha, n.var, fdr)
}
\arguments{
\item{mydata}{Data frame of data.}

\item{alpha}{Cutoff value of FDR or correlation.}

\item{n.var}{Number of elements in data.}

\item{fdr}{True/False - whether FDR to be used.}
}
\value{
a list of possible parents for each variable.
}
\description{
Find possible parents for each variable based on FDR or correlation curoff.
}
