% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{mypp1}
\alias{mypp1}
\title{Find possible parents for each variable based on FDR or correlation
cutoff for phenotype driven two-level possible parents search.}
\usage{
mypp1(mydata, alpha1, alpha2, n.var, y, fdr)
}
\arguments{
\item{mydata}{Data frame of data.}

\item{alpha1}{Cutoff value of FDR or correlation for second-level parents for outcome variable.}

\item{n.var}{Number of elements in data.}

\item{y}{Outcome variable.}

\item{fdr}{True/False - whether FDR to be used.}
}
\value{
a list of possible parents for each variable.
}
\description{
Find possible parents for each variable based on FDR or correlation
cutoff for phenotype driven two-level possible parents search.
}
