% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netword-plotting.R
\name{netplot_jm}
\alias{netplot_jm}
\title{plot network}
\usage{
netplot_jm(mylinks)
}
\arguments{
\item{mylinks}{data.frame with 3 columns; from, to and component}
}
\value{
a plot of the network
}
\description{
plot network
}
