% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{pofun}
\alias{pofun}
\title{Reformat a list of possible parents, pp, into a list of possible offspring, po}
\usage{
pofun(pp)
}
\arguments{
\item{pp}{List of possible parents.}
}
\value{
a list of possible offsprings for each variable.
}
\description{
Reformat a list of possible parents, pp, into a list of possible offspring, po
}
