% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{pp.sets}
\alias{pp.sets}
\title{Create all subsets of possible parents subsets for each node.}
\usage{
pp.sets(pp)
}
\arguments{
\item{pp}{List of possible parents.}
}
\value{
A list of all subsets of possible parents subsets for each node.
}
\description{
Create all subsets of possible parents subsets for each node.
}
