% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{pp.sets.s}
\alias{pp.sets.s}
\title{Compute the scores for all sets of possible parents for each node.}
\usage{
pp.sets.s(mydata, pps, surv, scoreFn)
}
\arguments{
\item{mydata}{Data frame of data with last column as output variable in case of binary/continuous outcome;
in case of survival data, data without the survival outcome and time.}

\item{pps}{List of possible parents' subsets.}

\item{surv}{Whether Survival outcome data, 0 or 1.}

\item{scoreFn}{Score Function to be used. Values 'bic' or 'bge'.}
}
\value{
List of scores for all sets of possible parents for each node.
}
\description{
Compute the scores for all sets of possible parents for each node.
}
