% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_all_best_networks.R
\name{return_all_best_networks}
\alias{return_all_best_networks}
\title{Return all best networks}
\usage{
return_all_best_networks(input)
}
\arguments{
\item{input}{Result from [caunset] function}
}
\value{
list of best networks
}
\description{
Return all best networks
}
\examples{
set.seed(1234)
mydata <- simdat(n_var = 10)
out <- causnet(data = mydata, max_parents = 3)

return_all_best_networks(out)
}
