% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_all_best_networks.R
\name{return_random_best_networks}
\alias{return_random_best_networks}
\title{Return n random best networks}
\usage{
return_random_best_networks(input, n)
}
\arguments{
\item{input}{Result from [caunset] function}

\item{n}{numeric}
}
\value{
list of best networks
}
\description{
Return n random best networks
}
\examples{
set.seed(1234)
mydata <- simdat(n_var = 10)
out <- causnet(data = mydata, max_parents = 3)

return_random_best_networks(out, 1)
}
