% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{score.bic.lm}
\alias{score.bic.lm}
\title{Bigger is better score for node y and parents x}
\usage{
score.bic.lm(y, x, mydat)
}
\description{
Bigger is better score for node y and parents x
}
\keyword{internal}
