% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{score.bic.surv}
\alias{score.bic.surv}
\title{Compute the BIC score of the survival outcome node with given set of parents.}
\usage{
score.bic.surv(y, x, mydata)
}
\arguments{
\item{y}{Child node.}

\item{x}{List of parents of y.}
}
\value{
BIC score.
}
\description{
Compute the BIC score of the survival outcome node with given set of parents.
}
