% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{score_bge}
\alias{score_bge}
\title{BGE scoring function}
\usage{
score_bge(y, x, mydat)
}
\arguments{
\item{y}{y}

\item{x}{x}

\item{mydat}{data}
}
\value{
numeric score
}
\description{
BGE scoring function
}
