% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_bic_lm.R
\name{score_bic_lm}
\alias{score_bic_lm}
\title{Bigger is better score for node y and parents x}
\usage{
score_bic_lm(y, x, mydat)
}
\description{
Bigger is better score for node y and parents x
}
\keyword{internal}
