% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdat.R
\name{simdat}
\alias{simdat}
\title{Simulate example data}
\usage{
simdat(sample.size = 300, n.var = 5, beta = 0.3)
}
\arguments{
\item{sample.size}{integer, number of observations}

\item{n.var}{integer, number of variables

The dependencies will be created in the fashion: 1 -> 2 -> 3 -> 4 -> 5}
}
\value{
data.frame
}
\description{
Simulate example data
}
