% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{sink2net}
\alias{sink2net}
\title{Find edges of a best network connected components from ordering of sinks.}
\usage{
sink2net(bnets, pp, pps, bps)
}
\arguments{
\item{bnets}{A list of best orderings of sinks.}

\item{pp}{List of Possible parents for all nodes.}

\item{pps}{List of Possible parents sets' subsets.}

\item{bps}{List of best parent subset in each Possible parents sets' subsets.}
}
\value{
A data frame of edges of a best network.
}
\description{
Find edges of a best network connected components from ordering of sinks.
}
