% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{subsetr}
\alias{subsetr}
\title{Compute index r for some set T, which is a subset of a set with n elements
watch out for rounding errors!!!}
\usage{
subsetr(n, set)
}
\arguments{
\item{n}{numeric}

\item{set}{set of elements}
}
\description{
Compute index r for some set T, which is a subset of a set with n elements
watch out for rounding errors!!!
}
\keyword{internal}
