% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{subsetur}
\alias{subsetur}
\title{Find a subset set T of a set with n elements, using index r
watch out for rounding errors!!!}
\usage{
subsetur(n, r)
}
\arguments{
\item{n}{numeric}

\item{r}{set of elements}
}
\description{
Find a subset set T of a set with n elements, using index r
watch out for rounding errors!!!
}
\keyword{internal}
