% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{swscore}
\alias{swscore}
\title{given a node s, and a set w, compute score for s with its best parent set in
w argument bps is generated by pp_sets_bs}
\usage{
swscore(s, w, pp, pps, bps)
}
\description{
given a node s, and a set w, compute score for s with its best parent set in
w argument bps is generated by pp_sets_bs
}
\keyword{internal}
