% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{wsink.scores}
\alias{wsink.scores}
\title{Given a set w of nodes, compute network scores for all possible sinks of sets, w1 = {w union v},
  where v is a possible offspring of at least one of the nodes in w.
W.networkscore is the score of the best network for w.}
\usage{
wsink.scores(w, w.networkscore, pp, po, pps, bps, m)
}
\description{
Given a set w of nodes, compute network scores for all possible sinks of sets, w1 = {w union v},
  where v is a possible offspring of at least one of the nodes in w.
W.networkscore is the score of the best network for w.
}
\keyword{internal}
