% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_sinks.R
\name{wsink_scores}
\alias{wsink_scores}
\title{Given a set w of nodes, compute network scores for all possible sinks of
sets, w1 = {w union v}, where v is a possible offspring of at least one of
the nodes in w. w_networkscore is the score of the best network for w.}
\usage{
wsink_scores(w, w_networkscore, pp, po, pps, bps, m)
}
\description{
Given a set w of nodes, compute network scores for all possible sinks of
sets, w1 = {w union v}, where v is a possible offspring of at least one of
the nodes in w. w_networkscore is the score of the best network for w.
}
\keyword{internal}
