% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causnet.R
\name{causnet}
\alias{causnet}
\title{Find globally optimal causal network}
\usage{
causnet(data, alpha = 0.05, max_parents = 2)
}
\arguments{
\item{data}{A data.frame of numeric variables.}

\item{alpha}{Numerical, p-value threshold to determine cutoff for possible
parents.}

\item{max_parents}{Numeric, maximal number of parents. Defaults to 2.}
}
\value{
A list with 2 elements: a data data.frame with 3 columns; from, to
and component, and a vector of number of best parents.
}
\description{
Find globally optimal causal network
}
\examples{
library(igraph)
set.seed(1)
new_data <- simdat(n_var = 5)
out <- causnet(data = new_data, 0.5)

netplot_jm(out)
}
