% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{score_bge}
\alias{score_bge}
\title{Compute the BGE score of a given child node with given set of parents.}
\usage{
score_bge(y, x, mydata)
}
\arguments{
\item{y}{Child node.}

\item{x}{List of parents of y.}
}
\value{
Numeric BGE score
}
\description{
Compute the BGE score of a given child node with given set of parents.
}
