% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfun.R
\name{sfun}
\alias{sfun}
\title{Main function to find best optimal BNs}
\usage{
sfun(
  mydata,
  surdata = NULL,
  scoreFn = "bic",
  pheno = FALSE,
  fdr = FALSE,
  alpha,
  alpha1 = 0.01,
  alpha2 = 0.01,
  pp = NULL,
  multBNs = FALSE
)
}
\arguments{
\item{mydata}{data.frame of data with last column as output variable
in case of binary/continuous outcome; in case of survival data, data without the survival outcome and time}

\item{surdata}{data.frame of data with only 2 columns as survival outcome and time, last col survival time}

\item{scoreFn}{type of scoring function to be used; options - "bic" and "bge"}

\item{pheno}{TRUE or False, for phenotype based search}

\item{alpha}{correlation cutoff for possible parent sets}

\item{alpha1}{First level correlation cutoff for possible parent sets in case of phenotype based search}

\item{alpha2}{Second level correlation cutoff for possible parent sets in case of phenotype based search}

\item{pp}{list of possible parent sets, if supplied by use; NULL otherwise}

\item{multBNs}{TRUE or False, for mutiple best networks}
}
\value{
(list of) data.frame(s) with 3 columns; from, to and component
}
\description{
Main function to find best optimal BNs
}
