% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{subsetr}
\alias{subsetr}
\title{Compute index r for some set T, which is a subset of a set with n elements}
\usage{
subsetr(n, T)
}
\arguments{
\item{n}{Number of elements in Superset of T.}

\item{T}{Subset of the Superset.}
}
\value{
index r for some set T.
}
\description{
Compute index r for some set T, which is a subset of a set with n elements
}
