% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{subsetur}
\alias{subsetur}
\title{Find a subset set T of a Superset with n elements, using index r}
\usage{
subsetur(n, r)
}
\arguments{
\item{n}{Number of elements in Superset of T.}

\item{r}{Index of the Subset of the Superset.}
}
\value{
a subset set T that has index r as a subset of Superset with n elements.
}
\description{
Find a subset set T of a Superset with n elements, using index r
}
