% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{swscore}
\alias{swscore}
\title{Given a node s, and a set w, compute score for s with its best parent set in w.
argument bps is generated by pp.sets.bs}
\usage{
swscore(s, w, pp, pps, bps)
}
\arguments{
\item{s}{node variable.}

\item{w}{Possible Parents subset for node s.}

\item{pps}{List of Possible parents sets' subsets.}

\item{bps}{List of best parent subset in each Possible parents sets' subsets.}
}
\value{
The score for s with its best parent set in w.
}
\description{
Given a node s, and a set w, compute score for s with its best parent set in w.
argument bps is generated by pp.sets.bs
}
