% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{wsink.scores}
\alias{wsink.scores}
\title{Given a set w of nodes, and v a possible offspring of at least one of the nodes in w,
compute network scores for all possible
sinks of sets, w1 = {w + v}.}
\usage{
wsink.scores(w, w.networkscore, pp, po, pps, bps, m)
}
\arguments{
\item{w}{A set of nodes.}

\item{w.networkscore}{The score of the best network for w.}

\item{pp}{List of possible parents.}

\item{po}{List of possible offsprings.}

\item{bps}{List of best parent subset in each Possible parents sets' subsets.}

\item{m}{Number of variables in the whole data.}
}
\value{
A list of network scores for all possible
sinks of setsof type w1 = {w + v}.
}
\description{
Given a set w of nodes, and v a possible offspring of at least one of the nodes in w,
compute network scores for all possible
sinks of sets, w1 = {w + v}.
}
